Remove the old Venafi issuer controller plugin
