+++ pkg/issuer/venafi/BUILD.bazel
@@ -1,26 +1,4 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "setup.go",
-        "venafi.go",
-    ],
-    importpath = "github.com/jetstack/cert-manager/pkg/issuer/venafi",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/api/util:go_default_library",
-        "//pkg/apis/certmanager/v1:go_default_library",
-        "//pkg/apis/meta/v1:go_default_library",
-        "//pkg/controller:go_default_library",
-        "//pkg/issuer:go_default_library",
-        "//pkg/issuer/venafi/client:go_default_library",
-        "//pkg/logs:go_default_library",
-        "@com_github_go_logr_logr//:go_default_library",
-        "@io_k8s_api//core/v1:go_default_library",
-        "@io_k8s_client_go//listers/core/v1:go_default_library",
-    ],
-)
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 filegroup(
     name = "package-srcs",
@@ -39,19 +17,15 @@ filegroup(
     visibility = ["//visibility:public"],
 )
 
-go_test(
-    name = "go_default_test",
-    srcs = ["setup_test.go"],
-    embed = [":go_default_library"],
+go_library(
+    name = "go_default_library",
+    srcs = ["legacy.go"],
+    importpath = "github.com/jetstack/cert-manager/pkg/issuer/venafi",
+    visibility = ["//visibility:public"],
     deps = [
+        "//pkg/api/util:go_default_library",
         "//pkg/apis/certmanager/v1:go_default_library",
         "//pkg/controller:go_default_library",
-        "//pkg/controller/test:go_default_library",
-        "//pkg/issuer/venafi/client:go_default_library",
-        "//pkg/issuer/venafi/client/fake:go_default_library",
-        "//pkg/logs:go_default_library",
-        "//pkg/util:go_default_library",
-        "//test/unit/gen:go_default_library",
-        "@io_k8s_client_go//listers/core/v1:go_default_library",
+        "//pkg/issuer:go_default_library",
     ],
 )
